add healthcheck endpoint, fix nuget package (antlr4) and Rubberduck.VBEditor issues; closes #21
