+++ .editorconfig
@@ -1,0 +1,4 @@
+﻿[*.cs]
+
+# IDE0057: Use range operator
+dotnet_diagnostic.IDE0057.severity = silent
