+++ Rubberduck.API/Program.cs
@@ -9,6 +9,8 @@ namespace Rubberduck.API
 #pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
     public class Program
     {
+        private static readonly string _rdapiPrefix = "rdapi_";
+
         public static void Main(string[] args)
         {
             CreateHostBuilder(args).Build().Run();
@@ -16,16 +18,22 @@ namespace Rubberduck.API
 
         public static IHostBuilder CreateHostBuilder(string[] args) =>
             Host.CreateDefaultBuilder(args)
+#if DEBUG == false // spares an ArgumentNullException at startup
                 .ConfigureAppConfiguration((context, config) =>
                 {
-                    var keyVaultEndpoint = new Uri(Environment.GetEnvironmentVariable("VaultUri"));
-                    config.AddAzureKeyVault(keyVaultEndpoint, new DefaultAzureCredential());
+                    try
+                    {
+                        var keyVaultEndpoint = new Uri(Environment.GetEnvironmentVariable("VaultUri"));
+                        config.AddAzureKeyVault(keyVaultEndpoint, new DefaultAzureCredential());
+                    }
+                    catch { }
                 })
+#endif
                 .ConfigureWebHostDefaults(webBuilder =>
                 {
                     var config = new ConfigurationBuilder()
                         .AddJsonFile("appsettings.json", optional: false)
-                        .AddEnvironmentVariables("rdapi_")
+                        .AddEnvironmentVariables(_rdapiPrefix)
                         .Build();
 
                     webBuilder.UseConfiguration(config);
