+++ Rubberduck.API/appsettings.json
@@ -7,7 +7,7 @@
     }
   },
   "ConnectionStrings": {
-    "RubberduckDb": "Data Source=(localdb)\\ProjectsV13;Initial Catalog=RDDB;Integrated Security=True;Persist Security Info=False;Pooling=False;MultipleActiveResultSets=False;Connect Timeout=60;Encrypt=False;TrustServerCertificate=False"
+    "RubberduckDb": "Data Source=(localdb)\\ProjectsV13;Initial Catalog=rubberduckdb;Integrated Security=True;Persist Security Info=False;Pooling=False;MultipleActiveResultSets=False;Connect Timeout=60;Encrypt=False;TrustServerCertificate=False"
   },
   "GitHub": {
     "UserAgent": "api.rubberduckvba.com",
