+++ RubberduckWebsite.UnitTests/RubberduckWebsite.UnitTests.csproj
@@ -1,0 +1,27 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+
+    <IsPackable>false</IsPackable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.9.4" />
+    <PackageReference Include="MSTest.TestAdapter" Version="2.2.3" />
+    <PackageReference Include="MSTest.TestFramework" Version="2.2.3" />
+    <PackageReference Include="coverlet.collector" Version="3.0.2" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Rubberduck.API\Rubberduck.API.csproj" />
+    <ProjectReference Include="..\RubberduckServices\RubberduckServices.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Reference Include="Rubberduck.SmartIndenter">
+      <HintPath>..\RubberduckServices\Libs\Rubberduck.SmartIndenter.dll</HintPath>
+    </Reference>
+  </ItemGroup>
+
+</Project>
