+++ RubberduckWebsite/RubberduckWebsite.csproj
@@ -5,6 +5,10 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <None Include="..\.editorconfig" Link=".editorconfig" />
+  </ItemGroup>
+
+  <ItemGroup>
     <ProjectReference Include="..\Rubberduck.Model\Rubberduck.Model.csproj" />
   </ItemGroup>
